Import all old blog posts.
These are all the published posts from the old blog, and all the
image files in the blog's gallery.

I started editing the first few posts to be markdowny, and use
newer links/local images/etc, but got bored. The rest are just
copies of the HTML, hotlinking files to the blog.

Add the smallest possible layouts to testit works.